Split files into .hpp .cpp main.cpp
