+++ bot/Makefile
@@ -1,0 +1,12 @@
+# **************************************************************************** #
+#                                                                              #
+#                                                         :::      ::::::::    #
+#    Makefile                                           :+:      :+:    :+:    #
+#                                                     +:+ +:+         +:+      #
+#    By: hania <hania@student.42.fr>                +#+  +:+       +#+         #
+#                                                 +#+#+#+#+#+   +#+            #
+#    Created: 2023/11/12 13:41:38 by hania             #+#    #+#              #
+#    Updated: 2023/11/12 13:41:39 by hania            ###   ########.fr        #
+#                                                                              #
+# **************************************************************************** #
+
